requirements.txt
Django>=4.0.2,<4.1.0
djangorestframework>=3.0.0,<3.14.0
psycopg2>=2.9.3<3.0.0
Pillow>=9.0.1,<9.1.0